Day4
Is using a parser combinator library *and* combination parsers overkill
for an AoC problem?  Well, yes, yes it is.  But that didn't stop me, now
did it? :) I thought it was kind of a fun exercise to do parsing and
input validation "in one go".  The code is not completely unreadable,
though I suppose doing this in two steps (though perhaps still only
exposing the final result to the user) is still nicer in general.

I have to say I thought it was pretty obvious that we would have to
do *something else* with the input for the second question (i.e. not
just "is this okay?") but, alas, that wasn't the caes.  The record might
then seem even more overkill than the use of a parser library for
something that's always just a simple Boolean answer.  One could argue
that it's more fun this way, so I'm going with that :)

Also yes, it's full of hacks; but that's the advent of code Christmas
spirit!