Day7
Fun with hash-maps again!  Also using ByteString instead of Text for
fun (and for O(1) dropEnd support :>).

Return of the Attoparsec, because parsing that line manually was not
very fun (I did that at first, before it occurred to me that I had
absolutely no fun writing it, so that needed to be fixed asap!)