Day13
Could've probably implemented the CRT myself, though why do that when
libraries have already made it fast? :)

What I do find a bit odd is that nowhere in the problem description was
it stated that the bus IDs are always coprime.  This is, of course, a
prerequisite for the CRT; leaving this out feels like the problem was
obscuring its solution just for the sake of it.