Day19
I'm pretty sure laziness, as well as the fact that the added rules are
not left-recursive, saved me here.  In practice this means that the
solution to part1 _just worked_ for part2, so yay Haskell \o/

In a strict language, you would probably have to manually inline the
rules that are not context free (this always terminates because all
strings are finite) and then work with that, or even just have some sort
of limit on how often you recurse.