Day22
It's a recursion!  Funnily enough, lists are actually faster here than
Data.Sequence sequences (for me, anyways); probably due to the latter's
rather high constant factors and the fact that list length is really not
the bottleneck in this case.  A contributing factor might also be that
Data.Sequence doesn't have a Hashable instance and hence one would have
to use Data.Set instead.