Day 7
The fact that one can finally use the mean for something (like helping
crabs) is pretty neat!

The average has to be computed with a little ±1 condition because it may
be that it's exactly `n.5`, in which case there's no obvious way to
decide which number will yield the lower value.  Funnily enough, my test
data ended up being pretty much exactly that (this always seems to good
to be an accident :).